fix(crypto): resolve 5 tsc errors in packages/crypto
## Root causes fixed

### 1. package.json — vitest y @prisma/client como devDeps
El test usaba `describe/it/expect` sin que vitest estuviera declarado.
`@prisma/client` faltaba para que ChannelType sea visible en tsc.

### 2. tsconfig.json — types vitest/globals + incluir __tests__
Sin types:['vitest/globals'], tsc no reconoce describe/it/expect
como globals (aunque vitest los inyecta en runtime).

### 3. credentials-schema.test.ts — eliminar import cross-package
import desde '../../../apps/api/...' viola el DAG del monorepo
(packages no puede depender de apps). El describe que usa
CreateChannelConfigSchema se mueve al test correcto: apps/api.
El test en crypto queda puro: solo valida lo que vive en @lss/crypto.

### 4. src/create-channel-config.schema.ts — schema movido a crypto
CreateChannelConfigSchema ahora vive en @lss/crypto y se re-exporta
desde apps/api importando '@lss/crypto'. Un solo source of truth.

### 5. apps/api DTO — ahora importa desde @lss/crypto
Eliminamos la definición duplicada. apps/api re-exporta
CreateChannelConfigSchema y CreateChannelConfigDto desde @lss/crypto.
