+++ packages/crypto/src/index.ts
@@ -1,5 +1,5 @@
 // packages/crypto/src/index.ts
-// @agent-vs/crypto — public API
+// @lss/crypto — public API
 
 export {
   encryptSecrets,
@@ -31,3 +31,7 @@ export type {
 } from './credentials-schema'
 
 export { encrypt, decrypt, encryptObject, decryptObject } from './aes'
+
+// CreateChannelConfigSchema — source of truth en @lss/crypto
+export { CreateChannelConfigSchema } from './create-channel-config.schema'
+export type { CreateChannelConfigDto } from './create-channel-config.schema'
